add details column to logs and use it to save question response text

+++ app/src/main/java/org/actionpath/db/logs/LogsDbHelper.java
@@ -15,17 +15,19 @@ public class LogsDbHelper extends SQLiteOpenHelper implements SyncableDbHelper {
     private static String TAG = LogsDbHelper.class.getName();
 
     private static final String DATABASE_NAME = "logs.db";
-    private static final int DATABASE_VERSION = 4;
+    private static final int DATABASE_VERSION = 5;
 
     // DB Table consts
     public static final String TABLE_NAME = "logs";
     public static final String ACTION_TYPE_COL = "actionType";
+    public static final String DETAILS_COL = "details";
 
     public static String[] LOGS_COLUMNS;
 
     static {
-        LOGS_COLUMNS = new String [] {
-                ID_COL, ACTION_TYPE_COL,
+        LOGS_COLUMNS = new String[]{
+                ID_COL,
+                ACTION_TYPE_COL, DETAILS_COL,  // these are unique to this table
                 INSTALLATION_ID_COL, ISSUE_ID_COL, TIMESTAMP_COL, LATITUDE_COL, LONGITUDE_COL, STATUS_COL
         };
     }
@@ -35,7 +37,8 @@ public class LogsDbHelper extends SQLiteOpenHelper implements SyncableDbHelper {
             + TABLE_NAME + "(" +
             ID_COL + " integer primary key autoincrement, " +
             ACTION_TYPE_COL + " text, " +
-            // and now the SyncableDbHelper columns
+            DETAILS_COL + " text, " +
+            // and now the  other SyncableDbHelper columns
             ISSUE_ID_COL + " int, " +
             INSTALLATION_ID_COL + " text, " +
             TIMESTAMP_COL + " timestamp, " +
@@ -58,9 +61,17 @@ public class LogsDbHelper extends SQLiteOpenHelper implements SyncableDbHelper {
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        Log.w(TAG, "Upgrading database from version " + oldVersion + " to " + newVersion + ", which will destroy all old data");
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
-        onCreate(db);
+        Log.w(TAG, "Upgrading database from version " + oldVersion + " to " + newVersion);
+        boolean upgraded = false;
+        if (oldVersion <= 4 && newVersion >= 5) {
+            db.execSQL("ALTER TABLE " + TABLE_NAME + " ADD COLUMN " + DETAILS_COL + " text;");
+            Log.i(TAG, "Upgraded " + TABLE_NAME + " from v1 to v2");
+            upgraded = true;
+        }
+        if (!upgraded) {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_NAME);
+            onCreate(db);
+        }
     }
 
 }
